refactor(llm): rename AnthropicModelLLM.ts to transitionLLM.ts
